Add row headers
Generate row multidimensional row headers from dimensions with the new
`createRowHeaders` function - check out its test to see the header row
data structure for different dimensions.
Change initial row dimensions to 2.
Use TitleRow when there is more than 1 header for a row.
Rename "Row" to "ValueRow" to make row for the functional "Row"
component that renders a variable number of title rows and one value
row, depending on the number of headers.

I've experimented with row-spanning headers, dropping the title rows,
but that did not work with the current layout. It would be an option for
the future.