Synchronize metadata, app state and sidebar
When the metadata loads, initialize the sidebar form fields with
dimension names in the right order.
Select the most-used X-Axis dimension when the metdaata loads and
synchronize that choice to the sidebar state, to ensure the sidebar
shows what's actually selected.

Move createDimensionCombinations to function, there is no need to attach
it to ScreenshotMetaData.