Change row header selection
Base the concept of headers on the amount of dimensions that are
displayed on the Y Axis: If there is only 1 dimension, it's displayed in
the "header column". If there are 2 dimensions, they are displayed as
horizontal "subheader rows". All following dimensions are displayed
underneath the images.

This gets rid of the old "contextInfo" concept and uses
selectedYSortOrder exclusively.