Improve typing and runtime checking of JSON data
Newer Typescript versions did no longer accept our definition of a
fixed-length array, needed to represent the dimension tuples in the JSON
metadata. The linter failed.

This commit uses the [runtypes](https://github.com/pelotom/runtypes)
library to define the data model of the metadata json and to validate
the data at runtime, so we're doing more than pseudo-typing.