Simplify row headers
Repetition of 202e5fbd07bfec950832d9931cfa69060f305f87 which was
accidentally pushed to the main branch.

This commit removes the concept of nested row headers. Originally
created for tables and row spans, then trying to display horizontal
headers, the algorithm in `createRowHeaders` that tried to build the
data structure for the headers was not working properly.

The new concept uses the first element of the Y-Axis sorting order array
as the header and displays all other information beneath each
screenshot.